ALWAYS_SEND_DURING_SPLIT:
this code always tries to send to the client if we're currently in a netsplit.
this should stop the dbufs from skyrocketing when we split.

please look over this and see if my logic is right. it hasn't been tested (therefore, it's not #defined).

If you have the chance, please look over the noquit stuff too. I haven't been able to test that on a scale
where things could really go wrong.


git-svn-id: svn+ssh://localhost/home/epiphani/home/dalrepo/bahamut/trunk@105 57140981-5af7-11db-9af8-080020d9fdd4