Fixed stupid archaic warning due.
Note to self: next time, commit the patch before you totally break channel.c
so no patches apply and you have to manually apply the patch.

Also fixed a problem with a get_client_name on an uninitialized variable on a
nick/server collision (not like this would ever happen, but consistancy is good).


git-svn-id: svn+ssh://localhost/home/epiphani/home/dalrepo/bahamut/trunk@124 57140981-5af7-11db-9af8-080020d9fdd4