Added a mini-block allocator to dbufs.
Previously we pre-allocated 1024 dbufs, and then malloc()'d each
dbuf after that one by one. Since each dbuf was 2060 bytes, this was
not very nice for the memory manager.

This also provides a way to implement a dbuf garbage collector, should we
ever have the need for one.

Removed all assertion calls, as a) they obviously have never happened
and b) they clutter the code.


git-svn-id: svn+ssh://localhost/home/epiphani/home/dalrepo/bahamut/trunk@130 57140981-5af7-11db-9af8-080020d9fdd4