Fixed "CAPAB" to something that is sent after PASS but before SERVER. Can only be sent once, can be sent by clients.
Fixed bug in sendto_noquit_servs_butone that wasn't uh.. doing things right.
(IsNoQuit(cptr) != 1 if it is true (it equals CAPAB_NOQUIT))

and the big meat of this patch...

Removal of the channel creation time from SJOIN and the channel struct.
Our 'channelts' is an accurate representation of this anyway.
Added new server<>server CAPAB, "SSJOIN" for Smart SJOIN:
  - backwards compatible (both ways)
  - new formats to servers also capable:
     serv<>serv sync/create: :server.in.us.dal.net SJOIN channelts #channel +stnk moo :@bob
     serv<>serv join: :bob SJOIN channelts #channel

  less bandwidth, same function, less cpu, tastes great, less filling!


git-svn-id: svn+ssh://localhost/home/epiphani/home/dalrepo/bahamut/trunk@141 57140981-5af7-11db-9af8-080020d9fdd4