This patch is fun:
New usermode +D, local umode. Users cannot accept DCCs of exploit-typed filenames
unless they specifically turn off +D, and +D is reset on each send of an exploit-type
filename. Informative notices are given. As a side effect, DCCs to channels are not
allowed.. this is good, as it will prevent annoying DCC chat floods that like
to crash certain irc clients.

Some may say this is a client side issue, but we simply cannot afford to be the victims
of the DDoS attack trojans that are spreading around like wildfire on our own network.
I've thought long and hard about this, and I feel this is the way to go. It really should
not have much effect on CPU usage at all, as DCCs aren't all that common in the grand
scheme of things.


git-svn-id: svn+ssh://localhost/home/epiphani/home/dalrepo/bahamut/trunk@319 57140981-5af7-11db-9af8-080020d9fdd4