Force the cached banserial of a user to be updated on a SVSNICK.
After services does a SVSNICK on a user, if the new nickname is banned
in a channel, they are still allowed to talk and change their nick if
they were not banned previously.

For example:

*** Mode change "+b testing!*@*" on #test by RuneB
*** Mode change "-o RuneB" on #test by RuneB
<RuneB> I can talk. Now I'll force a SVSNICK on me to testing.
*** RuneB is now known as testing
<testing> can I still talk?
<testing> yep.

This is because is_banned doesn't recheck the ban list if the banserial
has not changed for the user (instead it returns immediately). SVSNICK
forces a nick change, but doesn't change the banserial in any of the
channels the user is in.  This causes is_banned to think the ban status
has not changed.

Signed-off-by: epiphani <epiphani@dal.net>