Remove the use of ircsprintf.
Stub out the use of the ircsprintf.c file and use the native
sprintf/snprintf functions. Also tell GCC that various functions are
actually printf-like functions in disguise, and fix up the resulting
warnings.

This should help the code work better on 64-bit platforms, even if it
makes us a few microseconds slower.

Signed-off-by: epiphani <epiphani@dal.net>