Fix safe_ssl_read/safe_ssl_write to return -1 for EWOULDBLOCK/EAGAIN.
The safe_ssl_read and safe_ssl_write functions were setting errno and
then returning 0, which is incorrect. They should be returning -1
instead, since proper code only checks errno when -1 is returned from a
system call.

This patch should hopefully fix a bug that was causing SSL users to be
disconnected when a large number of bytes were being transferred from or
to the IRC server.

Signed-off-by: epiphani <epiphani@dal.net>