Introduce the possibility of multiple address families.
This patch adds an address family field to each client structure, which
is now checked in various places that expect to only be working with
IPv4 addresses, such as CIDR masking for /who and /rwho.

The IP address 0.0.0.1 is reserved to indicate that a non-IPv4 address
was passed through a server that does not have the NICKIPSTR capability.

Signed-off-by: epiphani <epiphani@dal.net>