Fix various problems with non-standard build configurations.
Libraries should be prepended to $LIBS instead of appended, as the
linker only does a single left-to-right pass over the command line.

This sometimes showed up as link errors (such as when shared libraries
are not being used) when -lssl came after -lcrypto, since -lssl depends
on -lcrypto.

Also, we were not actually using the include path for zlib detected by
configure, as it was setting $CPPFLAGS, which wasn't being used in the
Makefiles.

Signed-off-by: epiphani <epiphani@dal.net>