Clear dlerror() and ensure that dlsym never returns NULL
We were not clearing dlerror() before checking it, which caused us to
report the last error that occurred, even if it was not our fault.

Also check that dlsym never returns NULL. An error is better than a
crash in this case.

Original patch by Emilio Escobar, with some editing by me.

Signed-off-by: epiphani <epiphani@dal.net>