Fix identd connections on FreeBSD.
FreeBSD is apparently stricter than Linux with the address length
parameter to the connect system call. The code was using the total
length of the buffer (big enough for either an IPv4 or IPv6 address)
instead of the exact length depending on the kind of address, which
caused FreeBSD to return an error immediately, even though an IPv4
address can still fit in the buffer for an IPv6 address.

Fixed by re-using the address length calculated earlier when binding the
socket.

This was spotted thanks to a user who reported that identd wasn't
working for them.

Signed-off-by: epiphani <epiphani@dal.net>