Add the WEBIRC command.
This command is only allowed from trusted clients, and permits the real
hostname of a webirc client be used instead of the hostname of the
webirc gateway they connected through. This can be used to prevent users
from having ircip1.mibbit.com as their hostname, for example.

Trusted clients are added to the ircd.conf by adding special allow
blocks that match ident@host (if an identd response was received), or
webirc@host (if not), and that use a password of the form:
webirc.<password>

For example, in ircd.conf:

allow { ipmask webirc@127.0.0.1; passwd webirc.mypass; class users; };

will allow the command to be used from a non-identd client connecting
from 127.0.0.1 if they use the password mypass in the WEBIRC command.

Opers can also see if a client successfully used this command.

Signed-off-by: epiphani <epiphani@dal.net>