The following is for remote /rehash
Current m_rehash is now local_rehash.
The new m_rehash determines what to do with the rehash command it has
received, then either passes it to local_rehash or passes it on to the
target server.

To use remote rehash, the user on the local server must have access to
the rehash command on the server the user is on.

For the target server, the ircd.conf option "allow_remote_rehash" must
be enabled.

- Holbrook