Don't spam the network with the same channel topics
This patch changes m_topic() to ignore a new topic if the topic details (TS, topic and setter) are exactly the same.
It will prevent services from "flooding" the network with the same topics upon services restarts.
It won't affect users setting the same topic as the TS would be different.