Change "is now operator" to return umode instead of flag
Currently, this test is returning "O" for a network oper and "o" for a local oper, the opposite of what the umodes actually are. I presume that most people receiving this message will not have config file access, so it seems better to send the umode instead of the config flags.